+++ libs/base/utilities/src/misc/events-repeater.ts
@@ -1,3 +1,5 @@
+import { treewalk } from './treewalk';
+
 const eventsDataIdentifier = 'EVENTS_DATA';
 export const EVENTS_DATA = Symbol.for(eventsDataIdentifier);
 export const repeatableAttribute = 'repeatable';
@@ -9,41 +11,6 @@ export type ElementWithEventsData = HTMLElement & {
 interface EventsData {
   listener: (event: Event) => void;
   events: Event[];
-}
-
-function treewalk(
-  selector: string,
-  rootNode = document.body,
-  includeRoot = true
-) {
-  const nodes: Element[] = [rootNode];
-  const elements: HTMLElement[] = [];
-
-  for (const node of nodes) {
-    if (!node) continue;
-
-    if (node.nodeType !== Node.ELEMENT_NODE) {
-      continue;
-    }
-
-    if (node.children.length) {
-      nodes.push(...node.children);
-    }
-
-    if (node.shadowRoot?.children.length) {
-      nodes.push(...node.shadowRoot.children);
-    }
-
-    if (!includeRoot && node.matches(rootNode.tagName.toLowerCase())) {
-      continue;
-    }
-
-    if (node.matches(selector)) {
-      elements.push(node as HTMLElement);
-    }
-  }
-
-  return elements;
 }
 
 function eventsAction(
