+++ libs/platform/core/src/services/hydration/default-hydration.service.ts
@@ -7,6 +7,7 @@ import {
   HydratableLitElement,
   HYDRATE_ON_DEMAND,
   rootInjectable,
+  treewalk,
 } from '@spryker-oryx/utilities';
 import { LitElement } from 'lit';
 import { skip, Subscription, take } from 'rxjs';
@@ -50,7 +51,7 @@ export class DefaultHydrationService implements HydrationService, OnDestroy {
   }
 
   initHydrateHooks(immediate?: boolean): void {
-    const elementsToHydrate = this.treewalk(`[${hydratableAttribute}]`);
+    const elementsToHydrate = treewalk(`[${hydratableAttribute}]`);
 
     elementsToHydrate.forEach((el) => {
       if (immediate) {
@@ -84,11 +85,7 @@ export class DefaultHydrationService implements HydrationService, OnDestroy {
     }
 
     target.addEventListener(mode, async () => {
-      const childs = this.treewalk(
-        `[${deferHydrationAttribute}]`,
-        element,
-        false
-      );
+      const childs = treewalk(`[${deferHydrationAttribute}]`, element, false);
 
       if (childs.length) {
         const promises = childs.map((childEl) => this.hydrateOnDemand(childEl));
@@ -134,40 +131,5 @@ export class DefaultHydrationService implements HydrationService, OnDestroy {
 
   onDestroy(): void {
     this.subscription.unsubscribe();
-  }
-
-  protected treewalk(
-    selector: string,
-    rootNode = document.body,
-    includeRoot = true
-  ): HTMLElement[] {
-    const nodes: Element[] = [rootNode];
-    const elements: HTMLElement[] = [];
-
-    for (let i = 0; i < nodes.length; i++) {
-      const node = nodes[i];
-
-      if (node.nodeType !== Node.ELEMENT_NODE) {
-        continue;
-      }
-
-      if (node.children.length) {
-        nodes.push(...node.children);
-      }
-
-      if (node.shadowRoot?.children.length) {
-        nodes.push(...node.shadowRoot.children);
-      }
-
-      if (!includeRoot && node.matches(rootNode.tagName.toLowerCase())) {
-        continue;
-      }
-
-      if (node.matches(selector)) {
-        elements.push(node as HTMLElement);
-      }
-    }
-
-    return elements;
   }
 }
